Fix: serve the docs-markdown and blog-markdown index routes
/docs-markdown returned 400 "Missing 'path' query parameter" — there is no
such param; the empty slug was rejected before the lookup could reach
pages/docs/index.mdx. /blog-markdown 404'd because only a [slug] route
existed, so walking up from a post URL landed on nothing. Both are entry
points agents reach for first.

The blog index body is the listing /blog.txt already built, so that logic
moves to a shared blog-index module rather than being duplicated. /blog.txt
output is byte-identical.
